Conflicts:
	test/panda/core/RecvTest2.java
	test/panda/core/SenderTest2.java
	test/panda/tester/ReceiverTester.java

+++ test/panda/core/RecvTest2.java
@@ -14,7 +14,8 @@ public class RecvTest2
 	{
 		short topicID = Short.valueOf(args[0]).shortValue();
 		long numOfMessages = Long.valueOf(args[1]).longValue();
-		TopicInfo topicInfo = new TopicInfo("239.9.9.9", Integer.valueOf(9001), Short.valueOf(topicID), "TEST_TOPIC");
+		TopicInfo topicInfo = new TopicInfo("239.9.9.9", Integer.valueOf(9001), Integer.valueOf(topicID), "TEST_TOPIC");
+
 
 		ReceiverTester rt = new ReceiverTester();
 		rt.SubscribeToSequencedMessages(10000, topicInfo, numOfMessages, 10000000);
