+++ test/panda/core/SenderTest2.java
@@ -12,10 +12,11 @@ public class SenderTest2
 	 */
 	public static void main(String[] args) throws Exception
 	{
-		short topicID = Short.valueOf(args[0]).shortValue();
+		int topicID = Integer.valueOf(args[0]).intValue();
 		int payloadSize = Integer.valueOf(args[1]).intValue();
-		long numOfMessages = Long.valueOf(args[2]).longValue();
-		final TopicInfo topicInfo = new TopicInfo("239.9.9.9", Integer.valueOf(9001), Short.valueOf(topicID), "TEST_TOPIC");
+
+		int numOfMessages = Integer.valueOf(args[2]).intValue();
+		final TopicInfo topicInfo = new TopicInfo("239.9.9.9", Integer.valueOf(9001), Integer.valueOf(topicID), "TEST_TOPIC");
 
 		SenderTester st = new SenderTester();
 		// st.SendToTest(10000, topicInfo,
